refactor: move more code around
