+++ src/app/availability/[slug]/layout.tsx
@@ -1,4 +1,4 @@
-import { AvailabilityContextProvider } from "@/app/components/availability/availability-context";
+import { AvailabilityContextProvider } from "@/app/components/availability/context/availability-context";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
     return (
