+++ src/app/components/availability/availability-block-cell.tsx
@@ -1,24 +1,78 @@
 import { AvailabilityBlock } from "@/app/components/availability/availability-block";
 import { useAvailabilityContext } from "@/app/components/availability/context/availability-context";
+import { AvailabilityBlockType } from "@/lib/types/availability";
 import { cn } from "@/lib/utils";
+
+interface AvailabilityBlockCellProps {
+    blockIndex: number;
+    isAvailable: boolean;
+    zotDateIndex: number;
+    setAvailabilities: (startBlock: AvailabilityBlockType) => void;
+    isTopOfHour: boolean;
+    isHalfHour: boolean;
+    isLastRow: boolean;
+}
 
 export function AvailabilityBlockCell({
     blockIndex,
     isAvailable,
     zotDateIndex,
-    availabilitySelection,
-    handleTouchMove,
     setAvailabilities,
     isTopOfHour,
     isHalfHour,
     isLastRow,
-}) {
+}: AvailabilityBlockCellProps) {
     const {
         startBlockSelection,
         setStartBlockSelection,
         setEndBlockSelection,
         selectionState,
     } = useAvailabilityContext();
+
+    const availabilitySelection = {
+        zotDateIndex: zotDateIndex,
+        blockIndex: blockIndex,
+    };
+
+    const handleTouchMove = (e: React.TouchEvent) => {
+        const touchingElement = document.elementFromPoint(
+            e.touches[0].clientX,
+            e.touches[0].clientY
+        );
+
+        if (!touchingElement) return;
+
+        const touchingDateIndex = parseInt(
+            touchingElement.getAttribute("data-date-index") || "",
+            10
+        );
+        const touchingBlockIndex = parseInt(
+            touchingElement.getAttribute("data-block-index") || "",
+            10
+        );
+
+        if (
+            !isNaN(touchingDateIndex) &&
+            !isNaN(touchingBlockIndex) &&
+            startBlockSelection
+        ) {
+            setEndBlockSelection({
+                zotDateIndex: touchingDateIndex,
+                blockIndex: touchingBlockIndex,
+            });
+        }
+    };
+
+    const handleTouchEnd = (e: React.TouchEvent) => {
+        if (e.cancelable) {
+            e.preventDefault();
+        }
+
+        if (startBlockSelection) {
+            setEndBlockSelection(availabilitySelection);
+            setAvailabilities(startBlockSelection);
+        }
+    };
 
     const handleMouseUp = () => {
         if (startBlockSelection) {
@@ -46,17 +100,6 @@ export function AvailabilityBlockCell({
         setEndBlockSelection(availabilitySelection);
     };
 
-    const handleTouchEnd = (e: React.TouchEvent) => {
-        if (e.cancelable) {
-            e.preventDefault();
-        }
-
-        if (startBlockSelection) {
-            setEndBlockSelection(availabilitySelection);
-            setAvailabilities(startBlockSelection);
-        }
-    };
-
     return (
         <td
             onMouseUp={handleMouseUp}
@@ -65,9 +108,9 @@ export function AvailabilityBlockCell({
             <button
                 onTouchStart={handleTouchStart}
                 onTouchMove={handleTouchMove}
+                onTouchEnd={handleTouchEnd}
                 onMouseDown={handleMouseDown}
                 onMouseMove={handleMouseMove}
-                onTouchEnd={handleTouchEnd}
                 // TODO: Fix
                 // tabIndex="0"
                 data-date-index={zotDateIndex}
